test(datetime): migrate hour-cycle
